Consolidate more code
